refactor: extract OverviewTab component (274 lines)
- Extracted OverviewTab from page.tsx to components/OverviewTab.tsx
- Component displays key metrics, charts, and project health indicators
- Includes document status distribution pie chart and PMBOK health metrics
- File reduced from 3,094 to 2,829 lines (265 line reduction)
- Zero linter errors, functionality intact
